requirements.txt
dspy-ai>=2.6.0
transformers>=4.40.0
torch>=2.1.0
sentencepiece>=0.1.99
